refactor(sdk)!: move protocol message types to the protocol submodule (#1852)
