+++ sdk/packages/node/iii/package.json
@@ -71,6 +71,11 @@
       "types": "./dist/engine.d.ts",
       "import": "./dist/engine.mjs",
       "require": "./dist/engine.cjs"
+    },
+    "./protocol": {
+      "types": "./dist/protocol.d.ts",
+      "import": "./dist/protocol.mjs",
+      "require": "./dist/protocol.cjs"
     }
   },
   "dependencies": {
