+++ sdk/packages/node/iii/src/runtime.ts
@@ -1,1 +1,2 @@
+export type { IIIConnectionState } from './iii-constants'
 export type { FunctionRef, TriggerTypeRef } from './types'
