+++ sdk/packages/python/iii/src/iii/__init__.py
@@ -4,7 +4,6 @@ from typing import Any
 
 from .channels import ChannelReader, ChannelWriter
 from .errors import InvocationError
-from .format_utils import extract_request_format, extract_response_format, python_type_to_format
 from .iii import TriggerAction, register_worker
 from .iii_constants import (
     FunctionRef,
@@ -12,18 +11,9 @@ from .iii_constants import (
     TelemetryOptions,
 )
 from .iii_types import (
-    MessageType,
     MiddlewareFunctionInput,
-    RegisterFunctionFormat,
-    RegisterFunctionMessage,
-    RegisterTriggerInput,
-    RegisterTriggerMessage,
-    RegisterTriggerTypeInput,
-    RegisterTriggerTypeMessage,
     StreamChannelRef,
     TriggerActionEnqueue,
-    TriggerActionVoid,
-    TriggerRequest,
 )
 from .stream import IStream
 from .triggers import Trigger, TriggerConfig, TriggerHandler, TriggerTypeRef
@@ -49,17 +39,8 @@ __all__ = [
     # RBAC types
     "MiddlewareFunctionInput",
     # Message types
-    "MessageType",
-    "RegisterFunctionFormat",
-    "RegisterFunctionMessage",
-    "RegisterTriggerInput",
-    "RegisterTriggerMessage",
-    "RegisterTriggerTypeInput",
-    "RegisterTriggerTypeMessage",
     "StreamChannelRef",
     "TriggerActionEnqueue",
-    "TriggerActionVoid",
-    "TriggerRequest",
     # Triggers
     "Trigger",
     "TriggerConfig",
@@ -72,10 +53,6 @@ __all__ = [
     "StreamResponse",
     # Stream
     "IStream",
-    # Format extraction
-    "extract_request_format",
-    "extract_response_format",
-    "python_type_to_format",
 ]
