+++ sdk/packages/python/iii/src/iii/runtime.py
@@ -1,6 +1,6 @@
 """Public runtime types."""
 
-from .iii_constants import FunctionRef
+from .iii_constants import FunctionRef, IIIConnectionState
 from .triggers import TriggerTypeRef
 
-__all__ = ["FunctionRef", "TriggerTypeRef"]
+__all__ = ["FunctionRef", "IIIConnectionState", "TriggerTypeRef"]
