+++ sdk/packages/python/iii/src/iii/trigger.py
@@ -1,5 +1,6 @@
 """Public trigger types."""
 
+from .iii_types import TriggerActionVoid
 from .triggers import Trigger, TriggerConfig, TriggerHandler
 
-__all__ = ["Trigger", "TriggerConfig", "TriggerHandler"]
+__all__ = ["Trigger", "TriggerActionVoid", "TriggerConfig", "TriggerHandler"]
